requirements.txt
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.13.0
